Update of player points / Added Z-layaring for BGO - Player point should be removed from array if it not exist on map - you can set Z-position for each BGO, but only with LVLX format

+++ Editor/file_formats/file_formats.h
@@ -94,7 +94,7 @@ public:
     //PGE Extended File parameter string generators
     static QString IntS(long input);
     static QString BoolS(bool input);
-    static QString FloatS(float input);
+    static QString FloatS(double input);
     static QString qStrS(QString input);
     static QString hStrS(QString input);
     static QString strArrayS(QStringList input);
